Add startistics module and refactor
