+++ src/iact_estimator/core.py
@@ -15,6 +15,7 @@ from . import (
 )
 from .io import load_ebl
 from .spectral import crab_nebula_spectrum
+from .statistics import probability_to_sigma, sigma_to_probability, significance_li_ma
 
 __all__ = [
     "setup_logging",
@@ -22,7 +23,6 @@ __all__ = [
     "initialize_model",
     "observed_flux",
     "get_sed",
-    "significance_li_ma",
     "prepare_data",
     "source_detection",
     "calculate",
@@ -280,43 +280,7 @@ def get_sed(energy, flux):
     return sed
 
 
-def significance_li_ma(n_on, n_off, alpha, mu_sig=None):
-    """
-    Get the Li & Ma significance.
-
-    This is equivalent to eq.17 of [1]_.
-
-    Parameters
-    ----------
-    n_on : `int`
-        Measured counts in ON region.
-    n_off : `int`
-        Measured counts in OFF region.
-    alpha : `float`
-        Acceptance ratio of ON and OFF measurements.
-    mu_sig : `float`
-        Expected signal counts in ON region.
-
-    Returns
-    -------
-    sqrt_ts : `float``
-        Significance as the square root of the Test Statistic.
-
-    Notes
-    -----
-    The implementation uses `gammapy.stats.WStatCountsStatistic`
-    and takes the square root of the Test Statistic.
-
-    References
-    ----------
-    .. [1] Li, T.-P. & Ma, Y.-Q., ApJ, 1983, 272, 317, 10.1086/161295.
-    """
-    statistics = WStatCountsStatistic(n_on, n_off, alpha, mu_sig)
-    sqrt_ts = statistics.sqrt_ts
-    return sqrt_ts
-
-
-def prepare_data(config):
+def prepare_data(config, performance_data=None):
     """
     Extract the performance data.
